Add other_sub_type in figure
