+++ apps/event/enums.py
@@ -1,13 +1,10 @@
 __all__ = ['EventOtherSubTypeEnum']
 
 import graphene
-
-from apps.event.models import Event
-
 from utils.enums import enum_description
-from apps.common.enums import QA_RULE_TYPE
+from apps.common.enums import QA_RULE_TYPE, EVENT_OTHER_SUB_TYPE
 
-EventOtherSubTypeEnum = graphene.Enum.from_enum(Event.EVENT_OTHER_SUB_TYPE,
+EventOtherSubTypeEnum = graphene.Enum.from_enum(EVENT_OTHER_SUB_TYPE,
                                                 description=enum_description)
 QaRecommendedFigureEnum = graphene.Enum.from_enum(QA_RULE_TYPE, description=enum_description)
